code cleanup/refactor
