+++ src/lib/Components/Lists/ShowItemRow.tsx
@@ -6,10 +6,10 @@ import OpaqueImageView from "lib/Components/OpaqueImageView"
 import colors from "lib/data/colors"
 import { Pin } from "lib/Icons/Pin"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import { ExhibitionDates } from "lib/Scenes/Map/exhibitionPeriodParser"
 import { hrefForPartialShow } from "lib/utils/router"
 import { Schema, Track, track as _track } from "lib/utils/track"
 import { get } from "lodash"
-import moment from "moment"
 import React from "react"
 import { TouchableWithoutFeedback } from "react-native"
 import { commitMutation, createFragmentContainer, graphql, RelayProp } from "react-relay"
@@ -131,26 +131,6 @@ export class ShowItemRow extends React.Component<Props, State> {
     })
   }
 
-  getExhibitionPeriod() {
-    const {
-      show: { exhibition_period, status, end_at },
-    } = this.props
-    const twoYearsFromToday = moment()
-      .add(2, "years")
-      .utc()
-    const exhibitionEndDate = moment(end_at).utc()
-    const shouldDisplayOngoing = moment(exhibitionEndDate).isSameOrAfter(twoYearsFromToday)
-
-    if (shouldDisplayOngoing) {
-      return "Ongoing"
-    }
-    return (
-      status &&
-      exhibition_period &&
-      (status.includes("closed") ? status.charAt(0).toUpperCase() + status.slice(1) : exhibition_period)
-    )
-  }
-
   render() {
     const { noPadding, show, shouldHideSaveButton } = this.props
     const mainCoverImageURL = show.cover_image && show.cover_image.url
@@ -181,9 +161,14 @@ export class ShowItemRow extends React.Component<Props, State> {
                   {show.name}
                 </TightendSerif>
               )}
-              <Sans size="3t" color={color("black60")} ml={15}>
-                {this.getExhibitionPeriod()}
-              </Sans>
+              {show.exhibition_period &&
+                show.status && (
+                  <Sans size="3t" color={color("black60")} ml={15}>
+                    {show.status.includes("closed")
+                      ? show.status.charAt(0).toUpperCase() + show.status.slice(1)
+                      : ExhibitionDates(show.exhibition_period, show.end_at)}
+                  </Sans>
+                )}
             </Flex>
             {!shouldHideSaveButton && (
               <Flex flexDirection="row">
