+++ src/lib/Scenes/City/Components/Event/index.tsx
@@ -3,9 +3,9 @@ import { EventMutation } from "__generated__/EventMutation.graphql"
 import InvertedButton from "lib/Components/Buttons/InvertedButton"
 import OpaqueImageView from "lib/Components/OpaqueImageView"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import { ExhibitionDates } from "lib/Scenes/Map/exhibitionPeriodParser"
 import { Show } from "lib/Scenes/Map/types"
 import { Schema, Track, track as _track } from "lib/utils/track"
-import moment from "moment"
 import React from "react"
 import { TouchableWithoutFeedback } from "react-native"
 import { commitMutation, graphql, RelayProp } from "react-relay"
@@ -130,25 +130,9 @@ export class Event extends React.Component<Props, State> {
     SwitchBoard.presentNavigationViewController(this, `/show/${id}`)
   }
 
-  getExhibitionPeriod() {
-    const {
-      event: { exhibition_period, end_at },
-    } = this.props
-    const twoYearsFromToday = moment()
-      .add(2, "years")
-      .utc()
-    const exhibitionEndDate = moment(end_at).utc()
-    const shouldDisplayOngoing = moment(exhibitionEndDate).isSameOrAfter(twoYearsFromToday)
-
-    if (shouldDisplayOngoing) {
-      return "Ongoing"
-    }
-    return exhibition_period
-  }
-
   render() {
     const node = this.props.event
-    const { name, exhibition_period, partner, cover_image, is_followed } = node
+    const { name, exhibition_period, partner, cover_image, is_followed, end_at } = node
     const { name: partnerName } = partner
     const { isFollowedSaving } = this.state
     const url = cover_image ? cover_image.url : null
@@ -170,7 +154,7 @@ export class Event extends React.Component<Props, State> {
               </Serif>
               {exhibition_period && (
                 <Sans size="2" color={color("black60")}>
-                  {this.getExhibitionPeriod()}
+                  {ExhibitionDates(exhibition_period, end_at)}
                 </Sans>
               )}
             </TextContainer>
