+++ src/lib/Scenes/Show/Components/ShowHeader/ShowHeader.tsx
@@ -5,9 +5,9 @@ import InvertedButton from "lib/Components/Buttons/InvertedButton"
 import { EntityList } from "lib/Components/EntityList"
 import OpaqueImageView from "lib/Components/OpaqueImageView"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import { ExhibitionDates } from "lib/Scenes/Map/exhibitionPeriodParser"
 import { Schema, Track, track as _track } from "lib/utils/track"
 import { uniq } from "lodash"
-import moment from "moment"
 import React from "react"
 import { TouchableWithoutFeedback } from "react-native"
 import { commitMutation, createFragmentContainer, graphql, RelayProp } from "react-relay"
@@ -132,26 +132,10 @@ export class ShowHeader extends React.Component<Props, State> {
     SwitchBoard.presentNavigationViewController(this, `/show/${this.props.show.id}/artists`)
   }
 
-  getExhibitionPeriod() {
-    const {
-      show: { exhibition_period, end_at },
-    } = this.props
-    const twoYearsFromToday = moment()
-      .add(2, "years")
-      .utc()
-    const exhibitionEndDate = moment(end_at).utc()
-    const shouldDisplayOngoing = moment(exhibitionEndDate).isSameOrAfter(twoYearsFromToday)
-
-    if (shouldDisplayOngoing) {
-      return "Ongoing"
-    }
-    return exhibition_period
-  }
-
   render() {
     const { isFollowedSaving } = this.state
     const {
-      show: { artists, images, is_followed, name, partner, followedArtists },
+      show: { artists, images, is_followed, name, partner, followedArtists, end_at, exhibition_period },
     } = this.props
     const fairfollowedArtistList =
       (followedArtists && followedArtists.edges && followedArtists.edges.map(fa => fa.node.artist)) || []
@@ -171,7 +155,7 @@ export class ShowHeader extends React.Component<Props, State> {
           <Serif size="8" lineHeight={34}>
             {name}
           </Serif>
-          <Sans size="3">{this.getExhibitionPeriod()}</Sans>
+          {exhibition_period && <Sans size="3">{ExhibitionDates(exhibition_period, end_at)}</Sans>}
         </Box>
         {hasImages &&
           !singleImage && (
